Missed commits from shandong
