+++ packages/client/lib/rpc/helpers.ts
@@ -1,0 +1,32 @@
+import { bigIntToHex, bufferToHex, intToHex } from '@ethereumjs/util'
+
+import type { Block } from '@ethereumjs/block'
+import type { JsonRpcTx, TypedTransaction } from '@ethereumjs/tx'
+
+/**
+ * Returns tx formatted to the standard JSON-RPC fields
+ */
+export const jsonRpcTx = (tx: TypedTransaction, block?: Block, txIndex?: number): JsonRpcTx => {
+  const txJSON = tx.toJSON()
+  return {
+    blockHash: block ? bufferToHex(block.hash()) : null,
+    blockNumber: block ? bigIntToHex(block.header.number) : null,
+    from: tx.getSenderAddress().toString(),
+    gas: txJSON.gasLimit!,
+    gasPrice: txJSON.gasPrice ?? txJSON.maxFeePerGas!,
+    maxFeePerGas: txJSON.maxFeePerGas,
+    maxPriorityFeePerGas: txJSON.maxPriorityFeePerGas,
+    type: intToHex(tx.type),
+    accessList: txJSON.accessList,
+    chainId: txJSON.chainId,
+    hash: bufferToHex(tx.hash()),
+    input: txJSON.data!,
+    nonce: txJSON.nonce!,
+    to: tx.to?.toString() ?? null,
+    transactionIndex: txIndex !== undefined ? intToHex(txIndex) : null,
+    value: txJSON.value!,
+    v: txJSON.v!,
+    r: txJSON.r!,
+    s: txJSON.s!,
+  }
+}
