+++ packages/client/lib/rpc/modules/txpool.ts
@@ -1,3 +1,4 @@
+import { jsonRpcTx } from '../helpers'
 import { middleware } from '../validation'
 
 import type { EthereumClient } from '../..'
@@ -30,8 +31,11 @@ export class TxPool {
   content(_params = []) {
     const pending = new Map()
     for (const pool of this._txpool.pool) {
-      const pendingForAcct = pool[1].map((e) => e.tx.toJSON())
-      if (pendingForAcct.length > 0) pending.set('0x' + pool[0], pendingForAcct)
+      const pendingForAcct = new Map<bigint, any>()
+      for (const tx of pool[1]) {
+        pendingForAcct.set(tx.tx.nonce, jsonRpcTx(tx.tx))
+      }
+      if (pendingForAcct.size > 0) pending.set('0x' + pool[0], Object.fromEntries(pendingForAcct))
     }
     return {
       pending: Object.fromEntries(pending),
